requirements.txt
flask
stripe
python-dotenv
pytesseract
pillow